pyproject.toml
[project]
name = "sj-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "YulMitz", email = "yulmitz753220@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "shioaji>=1.2.5",
]

[project.scripts]
main = "sj_trading:main"
test = "sj_trading.testing_flow:testing_stock_ordering"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
